# products/meteo-office-desktop/services/skillhub/openapi.yaml
openapi: 3.1.0
info:
  title: MeteoMate SkillHub API
  version: 1.5.0
  description: Self-hosted registry for MeteoMate Experts and Agent Skills packages.
servers:
  - url: http://127.0.0.1:8088
paths:
  /healthz:
    get:
      responses:
        '200': {description: Healthy}
  /v1/auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200': {description: Authenticated desktop session}
        '401': {description: Invalid credentials or disabled account}
        '429': {description: Too many failed login attempts}
  /v1/auth/logout:
    post:
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Session revoked}
  /v1/me:
    get:
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Current actor and managed user profile}
    patch:
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Profile updated}
  /v1/me/password:
    post:
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Password changed and sessions revoked}
  /v1/me/policy:
    get:
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Effective organization, role, and user policy for the current desktop profile
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EffectivePolicyContext'}
  /v1/admin/users:
    get:
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Managed users}
    post:
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateUserRequest'}
      responses:
        '201': {description: User created}
  /v1/admin/users/{id}:
    patch:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: User updated}
  /v1/admin/users/{id}/reset-password:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Password reset and sessions revoked}
  /v1/admin/users/{id}/revoke-sessions:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: All managed sessions for the user revoked}
  /v1/admin/sessions:
    get:
      security: [{bearerAuth: []}]
      parameters:
        - {name: userId, in: query, schema: {type: string}}
      responses:
        '200':
          description: Active managed-user sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/Session'}}
                  total: {type: integer}
  /v1/admin/sessions/{id}:
    delete:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Session revoked}
        '404': {description: Session not found}
  /v1/admin/audit:
    get:
      security: [{bearerAuth: []}]
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: action, in: query, schema: {type: string}}
        - {name: limit, in: query, schema: {type: integer, maximum: 500}}
      responses:
        '200':
          description: Newest matching audit events
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/AuditEvent'}}
                  total: {type: integer}
  /v1/admin/policies:
    get:
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Organization baseline and all role and user overrides}
  /v1/admin/policies/organization:
    put:
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PolicySettings'}
      responses:
        '200': {description: Organization policy updated}
  /v1/admin/policies/roles/{role}:
    parameters:
      - {name: role, in: path, required: true, schema: {type: string, enum: [viewer, publisher, admin]}}
    put:
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PolicyPatch'}
      responses:
        '200': {description: Role policy override updated}
    delete:
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Role policy override reset}
  /v1/admin/policies/users/{id}:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string}}
    put:
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PolicyPatch'}
      responses:
        '200': {description: User policy override updated}
    delete:
      security: [{bearerAuth: []}]
      responses:
        '200': {description: User policy override reset}
  /v1/admin/policies/effective/users/{id}:
    get:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Effective desktop policy preview for a user
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EffectivePolicyContext'}
  /v1/trust/keys:
    get:
      responses:
        '200': {description: Public package-signing keys}
  /v1/skills:
    get:
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: category, in: query, schema: {type: string}}
        - {name: includeDrafts, in: query, description: Include manageable drafts; publishers see their own Skills and administrators see all Skills, schema: {type: boolean}}
        - {name: limit, in: query, schema: {type: integer, maximum: 200}}
        - {name: offset, in: query, schema: {type: integer, minimum: 0}}
      responses:
        '200': {description: Search results}
  /v1/skills/{id}:
    get:
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Skill descriptor and visible versions}
        '404': {description: Not found}
    patch:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateSkillRequest'}
      responses:
        '200': {description: Skill metadata, visibility, or ownership updated}
        '403': {description: Publisher is not the owner, or ownership transfer was attempted without administrator access}
        '404': {description: Skill not found}
  /v1/skills/{id}/versions:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [package]
              properties:
                package: {type: string, format: binary}
                metadata: {type: string, description: JSON-encoded upload metadata}
      responses:
        '201': {description: Draft version created}
  /v1/skills/{id}/versions/{version}/publish:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {name: version, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Version published and signed}
        '202': {description: Version submitted for administrator review}
  /v1/skills/{id}/versions/{version}/reject:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {name: version, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note: {type: string}
      responses:
        '200': {description: Pending version returned to draft}
  /v1/skills/{id}/versions/{version}/deprecate:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {name: version, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Published version deprecated}
  /v1/skills/{id}/versions/{version}/download:
    get:
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {name: version, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: ZIP package
          content:
            application/zip: {schema: {type: string, format: binary}}
  /v1/experts:
    get:
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: includeInactive, in: query, description: Include the current user's drafts, disabled, and archived Experts, schema: {type: boolean}}
        - {name: limit, in: query, schema: {type: integer, maximum: 300}}
        - {name: offset, in: query, schema: {type: integer, minimum: 0}}
      responses:
        '200':
          description: Experts visible to the current actor
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/Expert'}}
                  total: {type: integer}
    post:
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExpertMutation'}
      responses:
        '201':
          description: Expert created with revision 1
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Expert'}
        '409': {description: Expert ID already exists}
  /v1/experts/{id}:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string}}
    get:
      responses:
        '200':
          description: Visible Expert
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Expert'}
        '404': {description: Expert not found or not visible}
    put:
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExpertMutation'}
      responses:
        '200':
          description: Expert updated and a new immutable revision created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Expert'}
        '409': {description: baseRevision does not match the current revision}
  /v1/experts/{id}/submit-review:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '202':
          description: Expert submitted for administrator review
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Expert'}
        '409': {description: Archived Expert cannot be submitted}
  /v1/experts/{id}/review:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision, baseRevision]
              properties:
                decision: {type: string, enum: [approve, reject]}
                note: {type: string, maxLength: 500}
                baseRevision: {type: integer, minimum: 1}
      responses:
        '200':
          description: Expert review completed and a new revision created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Expert'}
        '409': {description: Expert is not pending review or baseRevision is stale}
  /v1/experts/{id}/status:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status, baseRevision]
              properties:
                status: {type: string, enum: [enabled, disabled, archived]}
                baseRevision: {type: integer, minimum: 1}
      responses:
        '200':
          description: Expert lifecycle status updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Expert'}
        '409': {description: Expert is not approved for enablement or baseRevision is stale}
  /v1/experts/{id}/distribution:
    put:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - {$ref: '#/components/schemas/ExpertDistribution'}
                - type: object
                  required: [baseRevision]
                  properties:
                    baseRevision: {type: integer, minimum: 1}
      responses:
        '200':
          description: Expert staged distribution updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Expert'}
        '409': {description: baseRevision is stale}
  /v1/experts/{id}/rollback/{revision}:
    post:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {name: revision, in: path, required: true, schema: {type: integer, minimum: 1}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [baseRevision]
              properties:
                baseRevision: {type: integer, minimum: 1}
      responses:
        '200':
          description: Historical Expert snapshot restored as a new revision
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Expert'}
        '409': {description: Rollback target is not older or baseRevision is stale}
  /v1/experts/{id}/revisions:
    get:
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Immutable Expert revision history
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: {type: array, items: {$ref: '#/components/schemas/ExpertRevision'}}
                  total: {type: integer}
  /v1/experts/{id}/revisions/{revision}:
    get:
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
        - {name: revision, in: path, required: true, schema: {type: integer, minimum: 1}}
      responses:
        '200':
          description: Immutable Expert revision
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExpertRevision'}
        '404': {description: Expert revision not found or not visible}
  /v1/recommendations:
    get:
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: categories, in: query, schema: {type: string}}
        - {name: installedSkillIds, in: query, schema: {type: string}}
        - {name: connectorIds, in: query, schema: {type: string}}
      responses:
        '200': {description: Ranked recommendations}
  /v1/installations:
    get:
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Installations visible to the current actor}
    post:
      security: [{bearerAuth: []}]
      responses:
        '201': {description: Installation reported or refreshed}
  /v1/installations/{id}:
    delete:
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200': {description: Installation record removed}
  /v1/admin/installations/summary:
    get:
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Installation metrics, version distribution, upgrade readiness, and pending publication reviews}
components:
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: {type: string}
        password: {type: string, format: password}
        clientId: {type: string}
    CreateUserRequest:
      type: object
      required: [username, password, role]
      properties:
        username: {type: string, pattern: '^[a-z0-9][a-z0-9._-]{2,63}$'}
        displayName: {type: string}
        password: {type: string, format: password, minLength: 8}
        role: {type: string, enum: [viewer, publisher, admin]}
        orgId: {type: string}
        mustChangePassword: {type: boolean}
    UpdateSkillRequest:
      type: object
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        summary: {type: string, maxLength: 300}
        description: {type: string, maxLength: 4000}
        categories: {type: array, maxItems: 12, items: {type: string}}
        tags: {type: array, maxItems: 24, items: {type: string}}
        icon: {type: string, maxLength: 16}
        visibility: {type: string, enum: [private, organization, public]}
        ownerId: {type: string, description: Administrator-only transfer to an active publisher or administrator}
    ExpertSource:
      type: object
      required: [type]
      properties:
        type: {type: string, enum: [user, organization, system]}
        remoteId: {type: string}
    ExpertReview:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [not_required, not_submitted, pending, approved, rejected]}
        note: {type: string}
        submittedBy: {type: string}
        submittedAt: {type: string, format: date-time}
        reviewedBy: {type: string}
        reviewedAt: {type: string, format: date-time}
    ExpertDistribution:
      type: object
      required: [mode]
      properties:
        mode: {type: string, enum: [all, percentage, allowlist]}
        percentage: {type: integer, minimum: 1, maximum: 100}
        userIds:
          type: array
          uniqueItems: true
          items: {type: string, minLength: 1}
    Expert:
      type: object
      required: [apiVersion, kind, id, name, version, revision, source, status, visibility, instruction, review, distribution, createdAt, updatedAt]
      properties:
        apiVersion: {type: string, const: meteomate.ai/v1}
        kind: {type: string, const: Expert}
        id: {type: string, pattern: '^[a-z0-9][a-z0-9._-]{2,127}$'}
        name: {type: string, minLength: 1, maxLength: 120}
        version: {type: string, minLength: 1, maxLength: 64}
        revision: {type: integer, minimum: 1}
        source: {$ref: '#/components/schemas/ExpertSource'}
        status: {type: string, enum: [draft, enabled, disabled, archived]}
        visibility: {type: string, enum: [private, organization, public]}
        owner: {type: string}
        ownerId: {type: string}
        orgId: {type: string}
        category: {type: string}
        avatar: {type: string}
        description: {type: string}
        mission: {type: string}
        tags: {type: array, items: {type: string}}
        instruction: {type: string, minLength: 1, maxLength: 64000}
        methodology: {type: array, items: {type: string}}
        workflow: {type: array, items: {type: string}}
        limitations: {type: array, items: {type: string}}
        inputs: {type: array, items: {type: string}}
        outputs: {type: array, items: {type: string}}
        prompts: {type: array, items: {type: string}}
        requiredSkills: {type: array, items: {type: string}}
        recommendedSkills: {type: array, items: {type: string}}
        requiredConnectors: {type: array, items: {type: string}}
        recommendedConnectors: {type: array, items: {type: string}}
        toolSelections:
          type: object
          additionalProperties:
            type: array
            items: {type: string}
        permissionProfile: {type: string}
        defaultWorkMode: {type: string, enum: [ask, plan, execute]}
        modelPolicy: {type: string}
        review: {$ref: '#/components/schemas/ExpertReview'}
        distribution: {$ref: '#/components/schemas/ExpertDistribution'}
        inputSchema: {type: object, additionalProperties: true}
        outputSchema: {type: object, additionalProperties: true}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    ExpertMutation:
      allOf:
        - {$ref: '#/components/schemas/Expert'}
        - type: object
          properties:
            baseRevision: {type: integer, minimum: 0, description: Required for updates and used for optimistic concurrency control}
    ExpertRevision:
      type: object
      required: [expertId, revision, version, snapshot, createdBy, createdAt]
      properties:
        expertId: {type: string}
        revision: {type: integer, minimum: 1}
        version: {type: string}
        snapshot: {$ref: '#/components/schemas/Expert'}
        createdBy: {type: string}
        createdAt: {type: string, format: date-time}
    Session:
      type: object
      required: [id, userId, clientId, createdAt, lastSeenAt, expiresAt]
      properties:
        id: {type: string}
        userId: {type: string}
        clientId: {type: string}
        createdAt: {type: string, format: date-time}
        lastSeenAt: {type: string, format: date-time}
        expiresAt: {type: string, format: date-time}
    AuditEvent:
      type: object
      required: [time, action, target, actor, remote]
      properties:
        time: {type: string, format: date-time}
        action: {type: string}
        target: {type: string}
        actor:
          type: object
          properties:
            subject: {type: string}
            name: {type: string}
            role: {type: string}
            orgId: {type: string}
        remote: {type: string}
        detail: {type: object, additionalProperties: true}
    PolicySettings:
      type: object
      required: [defaultModel, allowedModels, defaultSkillIds, allowedSkillIds, allowedConnectorIds, defaultPermissionProfileId, allowedPermissionProfileIds, autoCompactThreshold, skillPublishMode]
      properties:
        defaultModel: {type: string, description: Provider and model ID separated by a slash}
        allowedModels: {type: array, maxItems: 128, items: {type: string}}
        autoCompactThreshold: {type: number, minimum: 0.5, maximum: 0.95, default: 0.8}
        defaultSkillIds: {type: array, maxItems: 128, items: {type: string}}
        allowedSkillIds: {type: array, maxItems: 128, items: {type: string}}
        allowedConnectorIds: {type: array, maxItems: 128, items: {type: string}}
        defaultPermissionProfileId: {type: string, enum: ['', analysis-readonly, artifact-approval, workspace-approval]}
        allowedPermissionProfileIds:
          type: array
          maxItems: 3
          items: {type: string, enum: [analysis-readonly, artifact-approval, workspace-approval]}
        skillPublishMode: {type: string, enum: [publisher_direct, admin_approval]}
    PolicyPatch:
      type: object
      description: Omitted fields inherit from the lower-precedence scope.
      properties:
        defaultModel: {type: string}
        allowedModels: {type: array, maxItems: 128, items: {type: string}}
        autoCompactThreshold: {type: number, minimum: 0.5, maximum: 0.95}
        defaultSkillIds: {type: array, maxItems: 128, items: {type: string}}
        allowedSkillIds: {type: array, maxItems: 128, items: {type: string}}
        allowedConnectorIds: {type: array, maxItems: 128, items: {type: string}}
        defaultPermissionProfileId: {type: string, enum: ['', analysis-readonly, artifact-approval, workspace-approval]}
        allowedPermissionProfileIds:
          type: array
          maxItems: 3
          items: {type: string, enum: [analysis-readonly, artifact-approval, workspace-approval]}
    EffectivePolicy:
      allOf:
        - {$ref: '#/components/schemas/PolicySettings'}
        - type: object
          properties:
            sources: {type: object, additionalProperties: {type: string}}
            revision: {type: integer}
            updatedAt: {type: string, format: date-time}
    EffectivePolicyContext:
      type: object
      required: [userId, role, orgId, defaultSpaceId, profileBindingId, policy]
      properties:
        userId: {type: string}
        role: {type: string}
        orgId: {type: string}
        defaultSpaceId: {type: string}
        profileBindingId: {type: string}
        policy: {$ref: '#/components/schemas/EffectivePolicy'}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer